Implemented distance sensor for the parking_lot
When the system boots up takes a distance and sets it as the gate
distance. When the gate is open, if the sensor gets a distance closer
than the gate distance, it stays open since it means there is a car.
Also tweaked the welcome lcd function to display gate distance.