Add flying fish implementation
MH Sensor Flying Fish Implementation has been done in order to be used
in car counting part of the parking lot project. Interrupt driven design
has been used since it is not known when the car will exit the parking
lot. ISR of flying fish sensor is carefully implemented due to the fact
that even though it is defined as rising edge triggerred, it enters the
ISR in the falling edge as well. Prevented that issue.